requirements.txt
# MCP library
mcp-python==0.1.4

# Google Generative AI
google-generativeai>=0.3.0

# Data validation
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0

# HTTP client
httpx>=0.25.0

# Async support
asyncio-mqtt>=0.16.0
websockets>=12.0

# LangChain and MCP Adapters
langchain-mcp-adapters
langgraph
langchain-google-genai

# Configuration
pyyaml>=6.0

# HTTP Math Server
fastapi>=0.100.0
uvicorn>=0.20.0
aiohttp>=3.8.0


# LangChain and MCP Adapters 
langchain-mcp-adapters
langchain-openai
langchain-google-genai
langchain-anthropic
langchain-community
langchain-core
langchain